backend/requirements.txt
Django==5.1.7
djangorestframework==3.15.2
django-cors-headers==4.6.0
python-dotenv==1.0.1
dj-database-url==2.3.0
psycopg[binary]==3.2.6
gunicorn==23.0.0
